feat: Implement registerEditFromPrompt command for PerplexityPilot extension
